requirements.txt
Django==4.0.4
djangorestframework==3.14.0
PyJWT==2.5.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
Pillow==9.1.0